use arch linux for the CI build
boost installation fails due to the issues with python and numpy.
Reading through the GH issues for "Boost component 'numpy' is missing
libraries. Try building boost with '-o boost:without_numpy'."  it seems
like this does not have a solution right now, and I /need/ this boost
component to be installed.