build intermediate GH package with docker image
Almost half of the CI run time is spent on the dependency installation with =conan= itself
taking  at least  four minutes  of the  run time.  Creating an  intermediate package  that
further steps are built upon will (hopefully)  allow to (1) considerably reduce the ration
of  repetitive (package  installation)  work compared  to  the actual  test  time and  (2)
download the package and run some of the CI steps manually.